Risk Items

A table of risk items and their key attributes for project risk analysis.

db_risk

Format

A data frame of risk items with the following attributes:

OBJECTID

Geodatabase unique identifier, numeric

RISK_NO

Risk item number unique identifier, character

RISKCATEGORY

The general category of the risk (domain: risk_category), character

FEATURE

The primary project feature impacted by this action, character

DISCIPLINE

The primary discipline of the item (domain: discipline), character

RISK_OPPORTUNITY_EVENT

The risk event that prompted the creation of this risk item, character

CONCERNS

Description of the risk item, character

RISK_STRATEGY

The strategy for managing the risk (domain: risk_strategy), character

COST_LIKELIHOOD

Likelihood of the risk impacting project cost (domain: cost_likelihood), character

COST_IMPACT

Potential impact on project cost (domain: cost_impact), character

COST_RISK

Level of risk to project cost (low, moderate, high), character

COST_IMPACT_PER_PROJECT

Range of impact on project cost (in dollars), character

CONDITIONAL_FORMATTING

Cost risk color coding: low = green, moderate = orange, high = red, character

SCHEDULE_LIKELIHOOD

Likelihood of the risk impacting project schedule (domain: schedule_likelihood), character

SCHEDULE_IMPACT

Potential impact on project schedule (domain: schedule_impact), character

SCHEDULE_RISK

Level of risk to project schedule (low, moderate, high), character

SCHEDULE_IMPACT_PER_PROJECT

Range of impact to project schedule (in months), character

CONDITIONAL_FORMATTING2

Schedule risk color coding: low = green, moderate = orange, high = red, character

EFFICACY_IMPACT

Potential impact on project effectiveness (domain: efficacy_impact), character

EFFICACY_LIKELIHOOD

Likelihood of the risk impacting project effectiveness (domain: efficacy_likelihood), character

EFFICACY_RISK

Level of risk to project effectiveness (low, moderate, high), character

CONDITIONAL_FORMATTING3

Efficacy risk color coding: low = green, moderate = orange, high = red, character

TECHNICAL_POC

Technical POC for the action, character

POC_REVIEW_DATE

The last date the technical POC reviewed the action item, POSIXct

TIMING_OF_DELAY

Timing of delay relative to construction (prior to or during), character

START_DATE

Primary start date of the risk item, POSIXct

END_DATE

Primary end date of the risk item, POSIXct

ACTIVE

Is the item currently active (Yes/No)? character

TREND

Is the importance of the risk item decreasing, stable, or increasing, character

IMPACT

Brief description of the expected impact of the decision, character

OUTCOME

Brief description of the expected outcome for the risk item, character

MANAGEMENT_STATUS

The current management status of the item (domain: management_status, Open/Closed), character

OWNER

The name of the owner of the item, character

ENG_LEVEL

Current engagement level of the item (domain: Engagement, Level, 1-5), character

LEVEL_DATE

The date of the current engagement level, POSIXct

TAGS

Keyword tags used to categorize risk items, character

LAST_EDITED_DATE

The date the item was last eddited, POSIXct

PDT_DISC_CONC

not used

PDT_DISC_LIKELI_IMPACT

not used

PDT_DISC_RISK_MANAG_STRATEGY

not used

ACTION_ITEMS

not used

LESSONS_LEARNED

not used

FK_TABLE_ID

not used

Source

https://egis-app.mvr.usace.army.mil/ords/f?p=131