Format the decompressed chunk as an array of the correct type

When a chunk is decompressed it is returned as a vector of raw bytes. This function uses the array metadata to select how to convert the bytes into the final datatype and then converts the resulting output into an array of the appropriate dimensions, including re-ordering if the original data is in row-major order.

Usage

.format_chunk(decompressed_chunk, metadata, alt_chunk_dim)

Arguments

decompressed_chunk

Raw vector holding the decompressed bytes for this chunk.

metadata

List produced by .read_array_metadata() holding the contents of the .zarray file.

alt_chunk_dim

The dimensions of the array that should be created from this chunk. Normally this will be the same as the chunk shape in metadata, but when dealing with edge chunks, which may overlap the true extent of the array, the returned array should be smaller than the chunk shape.

Value

An array containing the decompressed chunk values.

If the output is larger than the space remaining in destination array i.e. it contains the overflowing elements, these will be trimmed when the chunk is returned to read_data()